requirements.txt
streamlit>=1.28.0
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=8.3.0
pandas>=1.3.0